Introduce RRuntimeManager class
