+++ extensions/positron-r/src/commands.ts
@@ -5,7 +5,7 @@
 import * as vscode from 'vscode';
 import * as positron from 'positron';
 import { timeout } from './util';
-import { getRunningRRuntime } from './provider';
+import { getRunningRRuntime } from './runtime';
 import { getRPackageName } from './contexts';
 import { getRPackageTasks } from './tasks';
 import { randomUUID } from 'crypto';
