+++ extensions/positron-r/src/formatting.ts
@@ -7,7 +7,7 @@ import * as positron from 'positron';
 import * as path from 'path';
 import * as os from 'os';
 import * as fs from 'fs';
-import { getRunningRRuntime } from './provider';
+import { getRunningRRuntime } from './runtime';
 import { timeout } from './util';
 import { randomUUID } from 'crypto';
